places on map query & saving images in db without coordinates
