+++ components/Map/MapGL.tsx
@@ -163,7 +163,9 @@ const MapGL: FC<{
             >
               <a className="">
                 <img
-                  src={image.url}
+                  src={
+                    'https://images.unsplash.com/photo-1561457013-a8b23513739a?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1124&q=80'
+                  }
                   className="rounded-full w-8 h-8 object-cover"
                   alt="Picture on map"
                 />
