+++ lib/mutations/createPost.ts
@@ -28,8 +28,6 @@ const CreatePost = async ({
     description:"${description}",
     createdAt: "${new Date().getTime()}",
     photoDate: "${photoDate}",
-    longitude: ${longitude},
-    latitude: ${latitude},
     url: "https://images.unsplash.com/photo-1561457013-a8b23513739a?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1124&q=80"
   })
   MERGE (place:Place {
