+++ lib/queries/GetPlaceInfo.ts
