+++ src/graphql/geo.graphql
@@ -22,6 +22,5 @@ query MapPosts(
     id
     latitude
     longitude
-    url
   }
 }
